feat: Optimization - only rerender updated components (time and tally)
